♻️ HappyLearning: split CourseQueryService and CourseLearningService
- Extract course query operations to CourseQueryService
- Extract course learning operations to CourseLearningService
- HappyLearningService now acts as facade delegating to both services
- Fix PASS_CHECK type from float to int to match DTO types
- Improve skip signal handling to skip only current lesson, continue next

+++ OfficalPlugin/SwissKitJ-Plugin-HappyLearning/src/main/java/plugin/swisskit/hpl/service/LearningConstants.java
@@ -25,7 +25,7 @@ public class LearningConstants {
     public static final float EXIT_TIME_DEDUCTION = 0.99999f;
 
     // Completion check
-    public static final float PASS_CHECK = 1.0f;
+    public static final int PASS_CHECK = 1;
 
     // Default values
     public static final int DEFAULT_PASSED = 0;
